+++ MulliganMadness/Cards/MMCard.cs
@@ -1,4 +1,3 @@
-using CardChoiceSpawnUniqueCardPatch.CustomCategories;
 using UnboundLib.Cards;
 using UnityEngine;
 
@@ -6,11 +5,6 @@ namespace MulliganMadness.Cards
 {
     public abstract class MMCard : CustomCard
     {
-        public const string TakebacksiesOnlyCategoryName = "MulliganMadness_TakebacksiesOnly";
-
-        public static CardCategory TakebacksiesOnlyCategory =>
-            CustomCardCategories.instance.CardCategory(TakebacksiesOnlyCategoryName);
-
         public override void SetupCard(CardInfo cardInfo, Gun gun, ApplyCardStats cardStats, CharacterStatModifiers statModifiers)
         {
             cardInfo.allowMultiple = AllowMultiple;