+++ TabInfoShim/Utils/PlayerCardBar.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-namespace TabInfo.Utils
-{
-    public class PlayerCardBar : MonoBehaviour
-    {
-        public PlayerFrame playerFrame;
-    }
-}