+++ TabInfoShim/Utils/TabFrame.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-namespace TabInfo.Utils
-{
-    public class TabFrame : MonoBehaviour
-    {
-        public bool toggled;
-    }
-}